refactor: move the pdpPresentation extractors into util.ts
They are pure functions and belong beside the other extraction helpers. It also
makes them unit-testable: index.ts starts a server on import, so nothing in it
can be exercised from a test without spawning a process.
